+++ CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 ## NEXT RELEASE
  - Improve DEVELOPPING.md with better testing command.
  - Fix an issue that make the Template Whisperer impossible to change in a entity.
+ - Add possibility to chose which Suggestion are avaialble for selection, by field.
+ - Add more tests to improve stability.
 
 ## 0.2.0 (2017-04-18)
  - Refactoring Template Whisperer as ConfigEntity instead of ContentEntity.